requirements.txt
annotated-types==0.7.0
anyio==4.10.0
APScheduler==3.10.4
bcrypt==4.0.1
certifi==2025.8.3
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
ecdsa==0.19.1
fastapi==0.119.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
passlib==1.7.4
pyasn1==0.6.1
pyasynchat==1.0.4
pyasyncore==1.0.4
pycparser==2.23
pydantic==2.12.0
pydantic_core==2.41.1
pyftpdlib==2.0.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.20
python-telegram-bot==22.4
pytz==2025.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uvicorn==0.37.0
